todtype and todevice methods + tests + all tests to float32

+++ src/spflow/base/inference/general/nodes/leaves/parametric/bernoulli.py
@@ -57,7 +57,7 @@ def log_likelihood(
     dispatch_ctx = init_default_dispatch_context(dispatch_ctx)
 
     # initialize probabilities
-    probs = np.zeros((data.shape[0], 1))
+    probs = np.zeros((data.shape[0], 1), dtype=node.dtype)
 
     # select relevant data based on node's scope
     data = data[:, node.scope.query]
